Basics Complete
Finished The basics of the application.
It works, not yet optimized for customazibility or use as a Tool. Current setup is meant to be used as is, feature plan is to make it more customazable and add additional features.

This version also lacks any real tests so it may have unknown issues